+++ x/stakers/keeper/msg_server_claim_commission_rewards.go
@@ -1,0 +1,44 @@
+package keeper
+
+import (
+	"context"
+
+	"cosmossdk.io/errors"
+	"github.com/KYVENetwork/chain/util"
+	"github.com/KYVENetwork/chain/x/stakers/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	errorsTypes "github.com/cosmos/cosmos-sdk/types/errors"
+)
+
+// ClaimCommissionRewards ...
+func (k msgServer) ClaimCommissionRewards(goCtx context.Context, msg *types.MsgClaimCommissionRewards) (*types.MsgClaimCommissionRewardsResponse, error) {
+	// Unwrap context and attempt to fetch the pool.
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// Check if staker exists
+	staker, found := k.GetStaker(ctx, msg.Creator)
+	if !found {
+		return nil, errors.Wrapf(errorsTypes.ErrNotFound, types.ErrNoStaker.Error(), msg.Creator)
+	}
+
+	// Check if amount can be claimed
+	if staker.CommissionRewards < msg.Amount {
+		return nil, types.ErrNotEnoughRewards.Wrapf("%d > %d", msg.Amount, staker.CommissionRewards)
+	}
+
+	// send commission rewards from stakers module to claimer
+	if err := util.TransferFromModuleToAddress(k.bankKeeper, ctx, types.ModuleName, msg.Creator, msg.Amount); err != nil {
+		return nil, err
+	}
+
+	// calculate new commission rewards and save
+	staker.CommissionRewards -= msg.Amount
+	k.setStaker(ctx, staker)
+
+	_ = ctx.EventManager().EmitTypedEvent(&types.EventClaimCommissionRewards{
+		Staker: msg.Creator,
+		Amount: msg.Amount,
+	})
+
+	return &types.MsgClaimCommissionRewardsResponse{}, nil
+}